% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{MBR-class}
\alias{MBR}
\alias{MBR-class}
\title{MBR objects}
\description{
MBR: an S4 class for co-regulation analysis and inference of 'dual 
regulons'.
}
\details{
The MBR class is a container for results from the MBR methods. 
The class slots are used to store information of different transcriptional networks, 
regulator annotation, infered 'dual regulons' and parameters used in the analysis. 
All the information is stored in nine slots.
}
\section{Slots}{

\describe{
\item{\code{TNI1}}{a 'TNI' object created by the RTN package.}

\item{\code{TNI2}}{another 'TNI' object created by the RTN package.}

\item{\code{testedElementsTNI1}}{regulatory elements listed in the TNI1.}

\item{\code{testedElementsTNI2}}{regulatory elements listed in the TNI2.}

\item{\code{dualRegulons}}{all possible 'duals regulons' computed by 
\code{\link[RTNmotifs:mbr.association]{mbr.association}}}

\item{\code{results}}{a list, results from the MBR methods.}

\item{\code{para}}{a list, parameters used in the MBR methods.}

\item{\code{summary}}{a list, summary for 'para' and 'results'.}

\item{\code{status}}{a character vector specifying the status of the MBR object 
based on the available methods.}
}}
\section{Constructor}{

newMBR (..., TNI1=NULL, TNI2=NULL, testedElementsTNI1=character(),
testedElementsTNI2=character(), dualRegulons=character(), results=list(), 
para=list(), summary=list(), status=character()): Creates an MBR object.

\itemize{
\item{...} - {Input data required to create the transcriptional networks: 
'gexp', 'regulatoryElements1' and 'regulatoryElements2'}
\item {TNI1} - {NULL, or an 'TNI' object.}
\item {TNI2} - {NULL, or an 'TNI' object.}
\item {testedElementsTNI1} - {A character vector.}
\item {testedElementsTNI2} - {A character vector.}
\item {dualRegulons} - {A character vector.}
\item {results} - {A list.}
\item {para} - {A list}
\item {summary} - {A list.}
\item {status} - {A character vector.}
}

There are two constructors to create an MBR object, users can opt for one of the following: 
(1) \code{\link[RTNmotifs:mbr.preprocess]{mbr.preprocess}}; 
(2) \code{\link[RTNmotifs:tni2mbr.preprocess]{tni2mbr.preprocess}}.
\itemize{
\item {(1)}: {It is used to create an MBR object without any pre-computed transcriptional network.}
\item {(2)}: {It is used to create an MBR object using available transcriptional networks.} 
}
}

