% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBR-methods.R
\docType{methods}
\name{mbr.dpi.filter,MBR-method}
\alias{mbr.dpi.filter}
\alias{mbr.dpi.filter,MBR-method}
\title{A filter based on the Data Processing Inequality (DPI) algorithm.}
\usage{
\S4method{mbr.dpi.filter}{MBR}(object, verbose = TRUE, ...)
}
\arguments{
\item{object}{A processed object of class \linkS4class{MBR} evaluated by 
the methods
 \code{\link[RTNmotifs:mbr.permutation]{mbr.permutation}} and 
 \code{\link[RTNmotifs:mbr.bootstrap]{mbr.bootstrap}}.}

\item{verbose}{A single logical value specifying to display detailed 
messages (when verbose=TRUE) or not (when verbose=FALSE).}

\item{...}{Additional arguments passed to the 
\code{\link[RTN:tni.dpi.filter]{tni.dpi.filter}} function.}
}
\value{
An \linkS4class{MBR} object with two DPI-filtered mutual information 
matrices, one in each "TNI" slot.
}
\description{
This function takes an MBR object and computes two transcriptional networks 
filtered by the data processing inequality algorithm.
}
\examples{
data("dt4rtn", package = "RTN")
gexp <- dt4rtn$gexp
annot <- dt4rtn$gexpIDs
tfs1 <- dt4rtn$tfs[c("IRF8","IRF1","PRDM1","AFF3","E2F3")]
tfs2 <- dt4rtn$tfs[c("HCLS1","STAT4","STAT1","LMO4","ZNF552")]
##---mbr.preprocess
rmbr <- mbr.preprocess(gexp=gexp, regulatoryElements1 = tfs1, 
regulatoryElements2=tfs2, gexpIDs=annot)
##---mbr.permutation
rmbr <- mbr.permutation(rmbr, nPermutations=10)
##---mbr.bootstrap
rmbr <- mbr.bootstrap(rmbr, nBootstrap=10)
##---mbr.dpi.filter
rmbr <- mbr.dpi.filter(rmbr)

}

