% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{MBR-class}
\alias{MBR-class}
\title{MBR: an S4 class for co-regulation analysis and inference of 'dual regulons'.}
\description{
MBR: an S4 class for co-regulation analysis and inference of 'dual regulons'.
}
\section{Slots}{

\describe{
\item{\code{TNI1}}{a 'TNI' object created by the RTN package.}

\item{\code{TNI2}}{another 'TNI' object created by the RTN package.}

\item{\code{testedElementsTNI1}}{regulatory elements listed in the TNI1.}

\item{\code{testedElementsTNI2}}{regulatory elements listed in the TNI2.}

\item{\code{dualsRegulons}}{all possible 'duals regulons' computed by \code{\link[RTNmotifs:mbr.association]{mbr.association}}}

\item{\code{results}}{a list, results from the MBR methods.}

\item{\code{para}}{a list, parameters used in the MBR methods.}

\item{\code{summary}}{a list, summary for 'para' and 'results'.}

\item{\code{status}}{a character vector specifying the status of the MBR object based on the available methods.}
}}

